Changed: Containers work with uint32_t now instead of size_t

+++ include/gaia/containers/sarray.h
@@ -4,7 +4,7 @@
 
 namespace gaia {
 	namespace containers {
-		template <typename T, size_t N>
+		template <typename T, uint32_t N>
 		using sarray = containers::sarr<T, N>;
 	} // namespace containers
 } // namespace gaia
