+++ include/gaia/ecs/component_utils.h
@@ -26,7 +26,7 @@ namespace gaia {
 
 				const auto& cc = ComponentCache::Get();
 				ComponentMatcherHash::Type hash = cc.GetComponentInfo(componentIds[0]).matcherHash.hash;
-				for (size_t i = 1; i < infosSize; ++i)
+				for (uint32_t i = 1; i < infosSize; ++i)
 					hash = utils::combine_or(hash, cc.GetComponentInfo(componentIds[i]).matcherHash.hash);
 				return {hash};
 			}
@@ -41,7 +41,7 @@ namespace gaia {
 
 				const auto& cc = ComponentCache::Get();
 				ComponentLookupHash::Type hash = cc.GetComponentInfo(componentIds[0]).lookupHash.hash;
-				for (size_t i = 1; i < infosSize; ++i)
+				for (uint32_t i = 1; i < infosSize; ++i)
 					hash = utils::hash_combine(hash, cc.GetComponentInfo(componentIds[i]).lookupHash.hash);
 				return {hash};
 			}
